PIPELINE: fix issues from codeFactor
